+++ jest.config.js
@@ -18,6 +18,8 @@ module.exports = {
   ],
   "coverageThreshold": {
     "global": {
+      "branches": 100,
+      "functions": 100,
       "lines": 100,
       "statements": 100
     }