+++ src/run.ts
@@ -4,26 +4,41 @@
 
 import * as core from '@actions/core'
 import {Api} from './gen/Api'
+import {LIB_VERSION} from './version'
 import {validate as uuidValidate} from 'uuid'
-import {PaginatedParameterList, Project} from './gen/data-contracts'
+import {PaginatedParameterList, Project, ProjectsParametersListParams} from './gen/data-contracts'
 import {HttpResponse} from './gen/http-client'
 
-require('isomorphic-fetch')
+import fetch from 'isomorphic-fetch'
+
+const USER_AGENT = `configure-action/${LIB_VERSION}`
+
+export const configurefetch = (
+  url: RequestInfo,
+  /* istanbul ignore next */
+  {headers, ...options}: RequestInit = {}
+) => {
+  return fetch(url, {
+    headers: {
+      'User-Agent': USER_AGENT,
+      ...headers
+    },
+    ...options
+  })
+}
 
 type SecurityDataType = {apikey: string}
 
 export function api(): Api<SecurityDataType> {
   const api = new Api<SecurityDataType>({
     baseUrl: core.getInput('server') || 'https://api.cloudtruth.io',
-    customFetch: fetch,
+    customFetch: configurefetch,
     securityWorker: (securityData: SecurityDataType | null) => {
-      if (securityData) {
-        return {
-          headers: {
-            ['Authorization']: 'Api-Key ' + securityData.apikey
-          },
-          keepalive: true
-        }
+      return {
+        headers: {
+          ['Authorization']: 'Api-Key ' + securityData!.apikey
+        },
+        keepalive: true
       }
     }
   })
@@ -58,35 +73,7 @@ function inject(response: HttpResponse<PaginatedParameterList>): void {
   }
 }
 
-async function resolve_environment_id(environment_name_or_id: string, api: Api<SecurityDataType>): Promise<string> {
-  if (!environment_name_or_id) {
-    throw new Error(`No environment name or id was specified.`)
-  }
-
-  if (uuidValidate(environment_name_or_id)) {
-    // we look it up to make sure the id is good and we have permission to use it
-    try {
-      const response = await api.environmentsRetrieve(environment_name_or_id)
-      return response.data.id
-    } catch (error: HttpResponse<Project, any> | any) {
-      throw new Error(`Environment "${environment_name_or_id}": ${error.error.detail}`)
-    }
-  }
-
-  const response = await api.environmentsList({name: environment_name_or_id})
-  if (response.data.count == 1) {
-    const result = response.data.results!
-    return result[0].id
-  }
-
-  throw new Error(`Environment "${environment_name_or_id}": Not found.`)
-}
-
 async function resolve_project_id(project_name_or_id: string, api: Api<SecurityDataType>): Promise<string> {
-  if (!project_name_or_id) {
-    throw new Error(`No project name or id was specified.`)
-  }
-
   if (uuidValidate(project_name_or_id)) {
     // we look it up to make sure the id is good and we have permission to use it
     try {
@@ -108,15 +95,24 @@ async function resolve_project_id(project_name_or_id: string, api: Api<SecurityD
 export async function run(): Promise<void> {
   try {
     const client = api()
-    const project_id = await resolve_project_id(core.getInput('project'), client)
-    const environment_id = await resolve_environment_id(core.getInput('environment'), client)
+    const project_id = await resolve_project_id(core.getInput('project', {required: true}), client)
+    const environment = core.getInput('environment', {required: true})
+    const tag = core.getInput('tag') || undefined
 
     for (let page = 1; ; ++page) {
-      core.debug(`Requesting parameter values for project='${project_id}' environment='${environment_id}' page=${page}`)
+      core.debug(
+        `Requesting parameter values for project='${project_id}' environment='${environment}' tag='${tag}' page=${page}`
+      )
+      let page_size = undefined
+      if (process.env.TESTING_REST_API_PAGE_SIZE) {
+        page_size = parseInt(process.env.TESTING_REST_API_PAGE_SIZE)
+      }
       const response = await client.projectsParametersList({
         projectPk: project_id,
-        environment: environment_id,
-        page: page
+        environment: environment, // can be name or id
+        tag: tag,
+        page: page,
+        page_size: page_size
       })
       inject(response)
       if (response.data.next == null) {